Pixy Detections Fix
I'm about as bright as clay sometimes. I used the the signature number
as the array reference. Rather than the current detection number. The
max default number of detections has been updated to its proper value
(1000). Be wary of memory issues with this update. 

Added a new note to the README.md todo list. Thanks E$!